// client/src/app/modules/quotation/model/ActeDeposit.ts
export interface ActeDeposit {
  id: number;
  isReductionCapitalSocial: boolean;
  isTransformation: boolean;
  isFusionScission: boolean;
  isFusionTransfrontaliere: boolean;
  isApportPartiel: boolean;
  isCessation: boolean;
  isSocialPartsCession: boolean;
  isSocialPartsDonation: boolean;
  isMandateExtension: boolean;
  isFusion: boolean;
  isScission: boolean;
  isAnnualAccountRefusal: boolean;
  isBylawsAmendments: boolean;
  isOrder: boolean;
  isLiquidatorReport: boolean;
  isContratAppui: boolean;
  isInformationConjoint: boolean;
  isAffectationPatrimoine: boolean;
  isDelegationPower: boolean;
}